before_symlink.rb edited online with Bitbucket

+++ deploy/before_symlink.rb
@@ -9,6 +9,7 @@ if !node['vagrant']
         code <<-EOH
         yum install -y https://s3.amazonaws.com/ec2-downloads-windows/SSMAgent/latest/linux_amd64/amazon-ssm-agent.rpm
         EOH
+        not_if { ::File.exist?('/usr/bin/amazon-ssm-agent') }        
     end
 end
